# docs/swagger.yaml
openapi: 3.0.0
info:
  title: ConnectMe API
  description: API for ConnectMe Application
  version: 1.0.0
servers:
  - url: https://cupcakez.my.id/api/v1
    description: Optional server description, e.g. Main (production) server
  - url: http://localhost:8080/api/v1
    description: Server in local

tags:
  - name: auth
    description: Operations about user authentication
  - name: workers
    description: Operations about workers
  - name: order
    description: Operations about order
  - name: review
    description: Operations about review
paths:
  /auth/signup:
    post:
      tags:
        - auth
      summary: User register
      description: User register to the application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fullName:
                  type: string
                  example: "Indra"
                email:
                  type: string
                  example: "xxxx@gmail.xom"
                password:
                  type: string
                  example: "indrachan"
                phone:
                  type: string
                  example: +62xxxxxxxxxxx
      responses:
        "200":
          description: A JSON with a token
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: 4b6e61c5-c231-4ab9-979d-xxxxxxxxxxxx
                  message:
                    type: string
                    example: User created successfully
  /auth/signin:
    post:
      tags:
        - auth
      summary: User login
      description: Login to the application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: "xxxx@gmail.xom"
                password:
                  type: string
                  example: indrachan
      responses:
        "200":
          description: A JSON with a token
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User login successfully
                  token:
                    type: string
  /auth/verify:
    post:
      tags:
        - auth
      summary: Verify user
      description: Verify user to able access application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                  example: 4b6e61c5-c231-4ab9-979d-xxxxxxxxxxxx
                code:
                  type: string
                  example: 1234
      responses:
        "200":
          description: A JSON with a id
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: 4b6e61c5-c231-4ab9-979d-xxxxxxxxxxxx
                  message:
                    type: string
                    example: User verified successfully

  /workers:
    get:
      tags:
        - workers
      summary: Get all workers
      description: Get all workers
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
            example: Bearer token
      responses:
        "200":
          description: A JSON with workers
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: List workers
                  workers:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          example: 079ca082-37c0-44d2-a648-f83bf482c4a3
                        name:
                          type: string
                          example: John Doe
                        tag:
                          type: object
                          properties:
                            id:
                              type: integer
                              example: 1
                            tag:
                              type: string
                              example: Plumber
                            specialization:
                              type: string
                              example: Plumbing
                        lower_price:
                          type: integer
                          example: 150000
                        image:
                          type: string
                          example: https://arcudskzafkijqukfool.supabase.co/storage/v1/object/public/connect-me/professional-plumber.jpg
                        review:
                          type: object
                          properties:
                            rating:
                              type: number
                              format: float
                              example: 4
                            total_rating:
                              type: integer
                              example: 2
  /workers/{id}:
    get:
      tags:
        - workers
      summary: Get worker by id
      description: Get worker by id
      parameters:
        - name: id
          in: path
          required: true
          description: Worker id
          schema:
            type: string
            example: 4b6e61c5-c231-4ab9-979d-xxxxxxxxxxxx
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
            example: Bearer token
      responses:
        "200":
          description: A JSON with worker
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Detail worker
                  worker:
                    type: object
                    properties:
                      id:
                        type: string
                        example: 079ca082-37c0-44d2-a648-f83bf482c4a3
                      name:
                        type: string
                        example: John Doe
                      tag:
                        type: object
                        properties:
                          id:
                            type: integer
                            example: 1
                          tag:
                            type: string
                            example: Plumber
                          specialization:
                            type: string
                            example: Plumbing
                      description:
                        type: string
                        example: Experienced plumber with over 5 years of work.
                      work_experience:
                        type: integer
                        example: 5
                      lower_price:
                        type: integer
                        example: 150000
                      worker_services:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: integer
                              example: 5
                            service:
                              type: string
                              example: Leak Repair
                            price:
                              type: integer
                              example: 150000
                      work_hour:
                        type: array
                        items:
                          type: string
                          example: "09:00"
                      image:
                        type: string
                        example: https://arcudskzafkijqukfool.supabase.co/storage/v1/object/public/connect-me/professional-plumber.jpg
                      review:
                        type: object
                        properties:
                          rating:
                            type: number
                            format: float
                            example: 4
                          total_rating:
                            type: integer
                            example: 2
                          reviews_detail:
                            type: array
                            items:
                              type: object
                              properties:
                                id:
                                  type: integer
                                  example: 1
                                name:
                                  type: string
                                  example: Gede Indra Adi Brata
                                review:
                                  type: string
                                  example: He is very professional and fast in fixing the leak in my house. I will definitely call him again if I have another problem.
                                rating:
                                  type: number
                                  format: float
                                  example: 4.5
                                time_sent:
                                  type: string
                                  example: 4 menit lalu
                          total_review:
                            type: integer
                            example: 2

  /orders:
    post:
      tags:
        - order
      summary: Create order
      description: Create order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                workerID:
                  type: string
                  example: 079ca082-37c0-44d2-a648-f83bf482c4a3
                workerService:
                  type: array
                  items:
                    type: integer
                  example: [1, 2]
                orderAddress:
                  type: object
                  properties:
                    street:
                      type: string
                      example: Jl. Raya Kedungkandang
                    latitude:
                      type: number
                      format: float
                      example: -7.9676
                    longitude:
                      type: number
                      format: float
                      example: 112.6581
                    addressType:
                      type: string
                      example: rumah
                    detailAddress:
                      type: string
                      example: depan indomaret
                date:
                  type: string
                  example: 20 October 2021
                time:
                  type: string
                  example: 08:00
                payment:
                  type: object
                  properties:
                    paymentMethod:
                      type: string
                      example: Gopay,BCA Virtual Account, BNI Virtua Account, Permata Virtual Account, BRI Virtual Account, ShopeePay
                    promoCode:
                      type: string
                      example: INDRA123
      responses:
        "200":
          description: A JSON with order
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Order Created
                  transaction:
                    type: object
                    properties:
                      limitTransactionDate:
                        type: string
                        example: 2024-09-25 08:55:52
                      totalPrice:
                        type: integer
                        example: 315000
                      virtualAccount:
                        type: object
                        properties:
                          virtualAccountNumber:
                            type: string
                            example: 300584103335182325
                          bankName:
                            type: string
                            example: bri
                      eWallet:
                        type: object
                        properties:
                          actions:
                            type: array
                            items:
                              type: object
                              properties:
                                name:
                                  type: string
                                  example: generate-qr-code
                                url:
                                  type: string
                                  example: https://api.sandbox.midtrans.com/v2/gopay/7814435b-6313-451a-89f4-52cfe0caf951/qr-code
    get:
      tags:
        - order
      summary: Get all orders
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
            example: Bearer token
        - name: status
          in: query
          required: false
          schema:
            type: string
            example: on_going,completed,canceled
      description: Get all orders
      responses:
        "200":
          description: A JSON with orders
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Orders Fetched
                  orders:
                    type: array
                    items:
                      type: object
                      properties:
                        orderID:
                          type: string
                          example: CM-1727237477
                        statusOrder:
                          type: string
                          example: On Going
                        workerID:
                          type: string
                          example: 079ca082-37c0-44d2-a648-f83bf482c4a3
                        workerName:
                          type: string
                          example: John Doe
                        workerImage:
                          type: string
                          example: https://arcudskzafkijqukfool.supabase.co/storage/v1/object/public/connect-me/professional-plumber.jpg
                        tag:
                          type: object
                          properties:
                            id:
                              type: integer
                              example: 1
                            tag:
                              type: string
                              example: Plumber
                            specialization:
                              type: string
                              example: Plumbing
                        totalPrice:
                          type: integer
                          example: 315000
                        orderDate:
                          type: string
                          example: 20 October 2021
                        orderTime:
                          type: string
                          example: 08:00
  /orders/{id}:
    get:
      tags:
        - order
      summary: Get order by id
      description: Get order by id
      parameters:
        - name: id
          in: path
          required: true
          description: Order id
          schema:
            type: string
            example: CM-1727237477
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
            example: Bearer token
      responses:
        "200":
          description: A JSON with order
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Order Fetched
                  order:
                    type: object
                    properties:
                      orderID:
                        type: string
                        example: CM-1727237477
                      statusOrder:
                        type: string
                        example: On Going
                      workerID:
                        type: string
                        example: 079ca082-37c0-44d2-a648-f83bf482c4a3
                      workerName:
                        type: string
                        example: John Doe
                      workerImage:
                        type: string
                        example: https://arcudskzafkijqukfool.supabase.co/storage/v1/object/public/connect-me/professional-plumber.jpg
                      tag:
                        type: object
                        properties:
                          id:
                            type: integer
                            example: 1
                          tag:
                            type: string
                            example: Plumber
                          specialization:
                            type: string
                            example: Plumbing
                      transactionTime:
                        type: string
                        example: 25 September 2024 11:11
                      paymentMethod:
                        type: string
                        example: BRI Virtual Account
                      paymentStatus:
                        type: string
                        example: On Going
                      location:
                        type: string
                        example: Jl. Raya Kedungkandang
                      workerService:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: integer
                              example: 1
                            service:
                              type: string
                              example: Leak Repair
                            price:
                              type: integer
                              example: 150000
                      serviceFee:
                        type: integer
                        example: 500
    patch:
      tags:
        - order
      summary: Create order
      description: Create order
      parameters:
        - name: id
          in: path
          required: true
          description: Order id
          schema:
            type: string
            example: CM-1727237477
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  example: finished,canceled,on_going
      responses:
        "200":
          description: A Message Update Order
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Order Updated

  /reviews:
    post:
      tags:
        - review
      summary: Create review
      description: Create review
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
            example: Bearer token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                workerID:
                  type: string
                  example: 079ca082-37c0-44d2-a648-f83bf482c4a3
                review:
                  type: string
                  example: He is very professional and fast in fixing the leak in my house. I will definitely call him again if I have another problem.
                rating:
                  type: number
                  format: float
                  example: 4.5
      responses:
        "200":
          description: A JSON with review
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Review Created
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT